requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
SQLAlchemy==2.0.39
pydantic==2.10.6
python-multipart==0.0.20
pydantic_core==2.27.2
passlib==1.7.4
aiofiles==24.1.0
bcrypt==4.3.0
python-dotenv ==1.0.1
python-jose==3.4.0
psycopg2-binary==2.9.10
email-validator==2.2.0
aiosmtplib==4.0.0
cloudinary==1.44.0